backend/requirements.txt
Django<3
djongo==1.2.38
uwsgi
sqlparse==0.2.4
dnspython==2.0.0
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-gmailapi-backend==0.2.0
djangorestframework-simplejwt==4.6.0
requests==2.25.1
google-auth==1.11.3
django-recaptcha2==1.4.1
django-ipware==3.0.2
googlemaps==4.4.5
jsonschema==3.2.0
google-cloud-storage==1.30.0
google-cloud-core==1.3.0
better-profanity==0.6.1
django-ckeditor==6.0.0
Pillow==8.2.0
docutils
drf-yasg